SetSecret gas fee optimization

+++ contracts/ymntest.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.15;
 
 contract YmnContract {
-    uint16 secret;
+    uint256 secret;
 
     constructor(uint16 initialSecret) {
         secret = initialSecret;
     }
 
-    function getSecret() public view returns(uint16) {
+    function getSecret() public view returns(uint256) {
         return secret;
     }
